Move SessionsView to a separate file
